Try to fix the core_ext loading issue
