+++ config/initializers/core_ext.rb
@@ -1,1 +1,10 @@
-require "core_ext/object"
+class Object
+  AssertionFailed = Class.new(StandardError)
+
+  def assert(test, message = nil)
+    unless test
+      message ||= "Expected #{test.inspect} to be truthy."
+      raise AssertionFailed, message
+    end
+  end
+end
