--- lib/core_ext/object.rb
@@ -1,10 +1,0 @@
-class Object
-  AssertionFailed = Class.new(StandardError)
-
-  def assert(test, message = nil)
-    unless test
-      message ||= "Expected #{test.inspect} to be truthy."
-      raise AssertionFailed, message
-    end
-  end
-end
